+++ CHANGELOG.md
@@ -4,15 +4,20 @@
 **BREAKING CHANGES**
 - Renamed `EditableField` class to `EditableText`
 
+### EditableText
+- Only show error help label if there is an error
+- If validationResult is unknown, return empty string instead of throwing an error
+
 ### Select
 - Update this.props.options proptypes to be more specific
 - Show the label for the selected option instead of always showing the placeholder
 
-### EditableSelect class added
+### EditableSelect Component added
 - Just as with EditableText, developers can now utilize an EditableSelect, that shows a static text value and when clicked it will transform into a Select dropdown.
+- Default behavior: onClick it will load the open Select dropdown and focus on the input search if it exists. Escape will close the dropdown and revert back to the clickable text div
 
-Note: This may have slight UI behavioral changes in the future such as removing the X button and on Esc keypress from the user reverting to static text as well as having the select dropdown be open by default when the static text is clicked.
-
+### CSS
+- Added `.disabled` class so any DOM element can benefit from the `not-allowed` cursor
 
 ## [v0.3.0]
 > May 29, 2016